refactor: move clickhouse node execution into its own module
ClickHouse was the last node still running inside workflow_executor.py.
Moves the body into clickhouse_node.execute, matching the other 61 nodes.
